+++ awokado/utils.py
@@ -1,14 +1,12 @@
-import datetime
-import hashlib
 import json
 import logging
 import random
 import string
 import sys
 import traceback
-import uuid
+from dataclasses import dataclass
 from json import JSONDecodeError
-from typing import NamedTuple, List, Dict, Any
+from typing import List, Dict, Any, Type, Tuple, Callable
 
 import falcon
 from dynaconf import settings
@@ -16,12 +14,16 @@ from sqlalchemy import desc, asc
 
 from awokado.consts import DEFAULT_ACCESS_CONTROL_HEADERS
 from awokado.exceptions import BaseApiException, IdFieldMissingError, BadRequest
-from awokado.filter_parser import parse_filters
+from awokado.filter_parser import FilterItem
+
+if False:
+    from awokado.resource import BaseResource
 
 log = logging.getLogger("awokado")
 
 
-class AuthBundle(NamedTuple):
+@dataclass
+class AuthBundle:
     user_id: int
     auth_token: str
 
@@ -30,7 +32,7 @@ def rand_string(size=8, chars=string.ascii_uppercase + string.digits):
     return "".join(random.choice(chars) for _ in range(size))
 
 
-def get_sort_way(sort_route):
+def get_sort_way(sort_route) -> Tuple[str, Callable]:
     if sort_route.startswith("-"):
         sort_route = sort_route[1:]
 
@@ -61,20 +63,16 @@ def empty_response(resource, is_list=False):
         return {resource_name: []}
 
 
-def get_read_params(req, resource) -> dict:
-    """
-    :param req: falcon.Request
-    :param resource: api.resources.base.BaseResource
-    """
-
-    params = dict(
-        include=req.get_param_as_list("include"),
-        sort=req.get_param_as_list("sort"),
-        limit=req.get_param_as_int("limit"),
-        offset=req.get_param_as_int("offset"),
-    )
-
-    params["filters"] = parse_filters(req._params, resource)
+def get_read_params(
+    req: falcon.Request, resource: Type["BaseResource"]
+) -> dict:
+    params = {
+        "include": req.get_param_as_list("include"),
+        "sort": req.get_param_as_list("sort"),
+        "limit": req.get_param_as_int("limit"),
+        "offset": req.get_param_as_int("offset"),
+        "filters": FilterItem.parse(req._params, resource),
+    }
     return params
 
 
@@ -136,7 +134,7 @@ def api_exception_handler(error, req, resp, params):
     elif isinstance(error, falcon.HTTPNotFound):
         resp.status = "404 Not Found"
         resp.content_type = "application/json"
-        resp.body = falcon.json.dumps({"error": f"{req.path} not found"})
+        resp.body = json.dumps({"error": f"{req.path} not found"})
         resp.append_header("Vary", "Accept")
 
     else:
@@ -147,7 +145,7 @@ def api_exception_handler(error, req, resp, params):
         if settings.get("AWOKADO_DEBUG"):
 
             if hasattr(error, "to_dict"):
-                resp.body = falcon.json.dumps({"error": error.to_dict()})
+                resp.body = json.dumps({"error": error.to_dict()})
 
             elif hasattr(error, "to_json"):
                 json_data = error.to_json()
@@ -157,27 +155,18 @@ def api_exception_handler(error, req, resp, params):
                 except (TypeError, JSONDecodeError):
                     json_data = json_data
 
-                resp.body = falcon.json.dumps({"error": json_data})
+                resp.body = json.dumps({"error": json_data})
 
             else:
                 exc_data = "".join(traceback.format_exception(*sys.exc_info()))
-                resp.body = falcon.json.dumps({"error": exc_data})
+                resp.body = json.dumps({"error": exc_data})
 
         else:
-            resp.body = falcon.json.dumps({"error": resp.status})
+            resp.body = json.dumps({"error": resp.status})
 
         # Set content type
         resp.content_type = "application/json"
         resp.append_header("Vary", "Accept")
-
-
-def get_uuid_hash(string):
-    return hashlib.md5(
-        (
-            str(uuid.uuid3(uuid.uuid4(), string))
-            + str(datetime.datetime.utcnow())
-        ).encode("utf-8")
-    ).hexdigest()
 
 
 def get_id_field(resource, name_only=False, skip_exc=False):